pyproject.toml
[project]
name = "shh_ai_pyton"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web core dependencies
    "aiofiles>=24.1.0",
    "celery==5.4.0",
    "db-libs>=1.1",
    "fastapi>=0.115.12",
    "funboost==48.4",
    "httpx==0.28.1",
    "orjson==3.10.15",
    "pandas==2.1.3",
    "psycopg2-binary==2.9.9",
    "redis==5.2.1",
    "sqlalchemy==2.0.38",
    "tqdm==4.67.1",
    "nibabel==5.3.2",
    "pydicom==2.4.3",
    "pydicom-seg>=0.4.1",
    "pyorthanc[all]>=1.20.0",
    "simpleitk==2.4.1",
    "matplotlib==3.8.2",
    "numpy==1.26.0",
    "numba==0.61.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "python-multipart>=0.0.20",
    "psutil>=7.0.0",
    "asyncpg>=0.30.0",
    "fastapi-cache2>=0.2.2",
    "reflex>=0.7.14",
    "advanced-alchemy>=1.4.4",
    "scikit-image>=0.25.2",
    "pynvml>=11.5.3",
    "openpyxl>=3.1.5",
]

[dependency-groups]
ai = [
    # AI and GPU related packages
    "brainextractor>=0.2.2",
    "nvidia-cublas-cu11==11.*",
    "nvidia-cuda-cupti-cu11==11.*",
    "nvidia-cuda-nvcc-cu11==11.*",
    "nvidia-cuda-nvrtc-cu11==11.*",
    "nvidia-cuda-runtime-cu11==11.8.*",
    "nvidia-cudnn-cu11==8.8.*",
    "nvidia-cufft-cu11>=10.9.0.58",
    "nvidia-curand-cu11==10.*",
    "nvidia-cusolver-cu11==11.*",
    "nvidia-cusparse-cu11==11.*",
    "nvidia-ml-py==11.*",
    "nvidia-nccl-cu11==2.*",
    "nvidia-nvtx-cu11==11.*",
    "opencv-python==4.7.0.72",
    "scikit-image==0.25.2",
    "scikit-learn>=1.6.1",
]

all = [
    { include-group = "ai" },
]
